requirements.txt
# Dev environment
pip
autopep8

# App
langchain_core
Pillow # PIL image-loading library
gradio